+++ Changelog.md
@@ -161,6 +161,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 
 ### Removed
 
+- Removed headshot mechanic, it was currently broken and will come back in the future
 - Removed all arcade guns (These will be moved to a sister mod at a later time)
 - Removed the SA80
 - Removed the mission system
@@ -193,6 +194,13 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 - [WIP] Renamed unintelligible variables to more readable names
 - Overhauled the internal process of item and block creation
 - Changed name and mod id
+
+### Balance Pack
+
+The new balance pack manager version is `1.3`. 
+
+- Added `inaccuracy`
+- Removed `headshotMultiplier`
 
 ##### Credits